Aplicando correcciones a comentarios
